Start hcm refactoring
