+++ bitbots_dynup/package.xml
@@ -38,8 +38,6 @@
   <test_depend>plotjuggler</test_depend>
   <test_depend>bitbots_odometry</test_depend>
 
-  <exec_depend>ros2launch</exec_depend>
-
   <export>
     <bitbots_documentation>
       <status>tested_integration</status>
