+++ bitbots_hcm/bitbots_hcm/hcm_dsd/actions/change_motor_power.py
@@ -16,10 +16,10 @@ class AbstractChangeMotorPower(AbstractActionElement):
         if not self.blackboard.visualization_active and not self.blackboard.simulation_active:
             # In visualization and simulation, we cannot disable motors
             try:
-                blackboard.motor_switch_service.wait_for_service(timeout_sec=10)
+                self.blackboard.motor_switch_service.wait_for_service(timeout_sec=10)
             except:
                 self.blackboard.node.get_logger().warn("HCM waiting for switch power service")
-            blackboard.motor_switch_service.wait_for_service()
+            self.blackboard.motor_switch_service.wait_for_service()
 
     def perform(self, reevaluate=False):
         raise NotImplementedError
